config-utils/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "config-utils"
version = "0.1.0"
description = "CLI tool for capturing environment variables and Django settings"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
django = ["django>=3.2"]

[project.scripts]
config-utils = "config_utils.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["config_utils"]
